Unify Net::HTTP and Faraday implementations
