+++ sentry-ruby/lib/sentry/faraday.rb
@@ -3,8 +3,6 @@
 module Sentry
   module Faraday
     OP_NAME = "http.client"
-    SPAN_ORIGIN = "auto.http.faraday"
-    BREADCRUMB_CATEGORY = "http"
 
     module Connection
       # Since there's no way to preconfigure Faraday connections and add our instrumentation
@@ -25,11 +23,10 @@ module Sentry
     end
 
     class Instrumenter
-      attr_reader :configuration
+      SPAN_ORIGIN = "auto.http.faraday"
+      BREADCRUMB_CATEGORY = "http"
 
-      def initialize
-        @configuration = Sentry.configuration
-      end
+      include Utils::HttpTracing
 
       def instrument(op_name, env, &block)
         return unless Sentry.initialized?
@@ -42,17 +39,14 @@ module Sentry
           end
 
           res = block.call
+          response_status = res.status
 
           if record_sentry_breadcrumb?
-            record_sentry_breadcrumb(request_info, res)
+            record_sentry_breadcrumb(request_info, response_status)
           end
 
           if sentry_span
-            sentry_span.set_description("#{request_info[:method]} #{request_info[:url]}")
-            sentry_span.set_data(Span::DataConventions::URL, request_info[:url])
-            sentry_span.set_data(Span::DataConventions::HTTP_METHOD, request_info[:method])
-            sentry_span.set_data(Span::DataConventions::HTTP_QUERY, request_info[:query]) if request_info[:query]
-            sentry_span.set_data(Span::DataConventions::HTTP_STATUS_CODE, res.status)
+            set_span_info(sentry_span, request_info, response_status)
           end
 
           res
@@ -65,40 +59,12 @@ module Sentry
         url = env[:url].scheme + "://" + env[:url].host + env[:url].path
         result = { method: env[:method].to_s.upcase, url: url }
 
-        if configuration.send_default_pii
+        if Sentry.configuration.send_default_pii
           result[:query] = env[:url].query
           result[:body] = env[:body]
         end
 
         result
-      end
-
-      def record_sentry_breadcrumb(request_info, res)
-        crumb = Sentry::Breadcrumb.new(
-          level: :info,
-          category: BREADCRUMB_CATEGORY,
-          type: :info,
-          data: {
-            status: res.status,
-            **request_info
-          }
-        )
-
-        Sentry.add_breadcrumb(crumb)
-      end
-
-      def record_sentry_breadcrumb?
-        configuration.breadcrumbs_logger.include?(:http_logger)
-      end
-
-      def propagate_trace?(url)
-        url &&
-          configuration.propagate_traces &&
-          configuration.trace_propagation_targets.any? { |target| url.match?(target) }
-      end
-
-      def set_propagation_headers(headers)
-        Sentry.get_trace_propagation_headers&.each { |k, v| headers[k] = v }
       end
     end
   end
